feature: UNSET, change export folder and .h name to env_management, leaks tested. NEED TO FIX: export (var) without '='
