+++ Makefile
@@ -1,5 +1,5 @@
 NAME=minishell
-FILES=./src/main.o ./src/tokenizer/wordbreaker.o ./src/tokenizer/lexer.o ./src/sentences/pipes.o ./src/tokenizer/parser.o ./src/env/get_env.o ./src/env/utils.o ./src/export/export.o ./src/export/utils.o
+FILES=./src/main.o ./src/tokenizer/wordbreaker.o ./src/tokenizer/lexer.o ./src/sentences/pipes.o ./src/tokenizer/parser.o ./src/env/get_env.o ./src/env/utils.o ./src/env_management/export.o ./src/env_management/unset.o ./src/env_management/utils.o
 
 LIBFT_PATH=./include/libft
 LIBFT_NAME=./include/libft/libft.a
