+++ src/minishell.h
@@ -17,7 +17,7 @@
 #include "./tokenizer/tokenizer.h"
 #include "./sentences/sentences.h"
 #include "./env/env.h"
-#include "./export/export.h"
+#include "./env_management/env_management.h"
 //#include <readline/readline.h>
 //#include <readline/history.h>
 //#include <term.h>
